refine graph python stack
