+++ tools/functional/generate_functional_api.py
@@ -122,11 +122,11 @@ pybind_source_fmt = (
 #include "oneflow/api/python/functional/function_def.h"
 #include "oneflow/api/python/functional/python_arg.h"
 #include "oneflow/api/python/functional/python_arg_parser.h"
-#include "oneflow/api/python/functional/python_frame.h"
 #include "oneflow/api/python/functional/python_return_types.h"
 #include "oneflow/core/common/maybe.h"
 #include "oneflow/core/common/optional.h"
 #include "oneflow/core/functional/functional.h"
+#include "oneflow/extension/stack/python/stack_getter.h"
 
 namespace {{
 // This return type template code is referenced from:
