+++ tools/functional/generate_tensor_api.py
@@ -118,11 +118,11 @@ pybind_source_fmt = (
 #include "oneflow/api/python/functional/function_def.h"
 #include "oneflow/api/python/functional/python_arg.h"
 #include "oneflow/api/python/functional/python_arg_parser.h"
-#include "oneflow/api/python/functional/python_frame.h"
 #include "oneflow/api/python/functional/tensor_api.yaml.h"
 #include "oneflow/api/python/functional/tensor_api.yaml.pybind.h"
 #include "oneflow/core/common/maybe.h"
 #include "oneflow/core/common/optional.h"
+#include "oneflow/extension/stack/python/stack_getter.h"
 
 namespace oneflow {{
 namespace one {{
